requirements.txt
attrs==21.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
flasgger==0.9.5
Flask==2.0.2
Flask-Cors==3.0.10
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.0
MarkupSafe==2.0.1
mistune==0.8.4
pyrsistent==0.18.0
PyYAML==6.0
six==1.16.0
typing-extensions==3.10.0.2
waitress==2.0.0
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.6.0
boto3~=1.18.21
msrest~=0.6.21
googletrans~=4.0.0rc1